Fixing siteHeader Directive
The siteHeader directive was never really done. It had problems. Replacing
the 'link' method with a isolate scope and a slightly more complicated
template that's working better.

Still looks ugly because I'm using Bootstrap styles but I'm not using
Bootstrap anymore, so I've got some CSS work to do to fix that up.